requirements.txt
locustio>=0.8a2
Faker
six
